# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Cria um JWT para o cliente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login cliente",
                "parameters": [
                    {
                        "description": "Dados do Cliente",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Credentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Customer"
                        }
                    }
                }
            }
        },
        "/customer": {
            "post": {
                "description": "Adicionar um novo cliente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer"
                ],
                "summary": "Adicionar novo cliente",
                "parameters": [
                    {
                        "description": "Dados do Cliente",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CustomerInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Customer"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Credentials": {
            "type": "object",
            "properties": {
                "cpf": {
                    "type": "string"
                },
                "pass": {
                    "type": "string",
                    "format": "string"
                }
            }
        },
        "entity.Customer": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "usuario@host.com"
                },
                "id": {
                    "type": "string",
                    "format": "string",
                    "example": "1"
                },
                "user": {
                    "type": "string",
                    "example": "paulo"
                }
            }
        },
        "entity.CustomerInput": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "usuario@host.com"
                },
                "id": {
                    "type": "string",
                    "format": "string",
                    "example": "1"
                },
                "pass": {
                    "type": "string",
                    "format": "string"
                },
                "user": {
                    "type": "string",
                    "example": "paulo"
                }
            }
        }
    }
}